# src/main/java/travel/rewardo/rewardapi/scraper/vs/model/api/AwardBusiness.java
package travel.rewardo.rewardapi.scraper.vs.model.api;

import jakarta.persistence.Entity;
import jakarta.persistence.GeneratedValue;
import jakarta.persistence.GenerationType;
import jakarta.persistence.Id;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
@Entity
public class AwardBusiness {
    @Id
    @GeneratedValue(strategy = GenerationType.UUID)
    private String id;
    private Integer cabinPointsValue;
    private Boolean isSaverAward;
    private Integer cabinClassSeatCount;
    private String cabinClassSeatCountString;
}