# src/main/java/travel/rewardo/rewardapi/scraper/vs/model/data/AwardBusiness.java
package travel.rewardo.rewardapi.scraper.vs.model.data;

import jakarta.persistence.*;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
@Entity(name = "REWARD_FLIGHT_LATEST_AWARD_BUSINESS")
@Table(name = "REWARD_FLIGHT_LATEST_AWARD_BUSINESS")
public class AwardBusiness {
    @Id
    @GeneratedValue(strategy = GenerationType.UUID)
    private String id;
    private Integer cabinPointsValue;
    private Boolean isSaverAward;
    private Integer cabinClassSeatCount;
    private String cabinClassSeatCountString;
}